feat: add governance infrastructure improvements for EU AI Act compliance
- Audit trail durability: AuditBackend Protocol with InMemoryAuditBackend
  and JSONLAuditBackend (append-only with fsync). AuditLog.from_backend()
  reconstructs chain from durable storage.

- Circuit breaker (Article 14 kill-switch): GovernanceCircuitBreaker with
  thread-safe in-process + cross-process file signaling. Wired into
  GovernedAgent.run() as Step 0 check.

- Compliance plugin system: setuptools entry point discovery for
  third-party frameworks via acgs_lite.compliance_frameworks group.
  register_framework(), register_jurisdiction(), register_domain() APIs.

- Runtime evidence auto-generation: AuditLogEvidenceCollector maps audit
  entry patterns to EU AI Act articles (Art.9, Art.12, Art.14). Chain
  integrity verification generates Art.12(2) evidence at 0.95 confidence.

- Rust 127-rule ceiling: u64→u128 bitmask in validator.rs, doubling max
  constitution size from 63 to 127 rules.

- enhanced_agent_bus decoupling: extracted Protocol types and lazy import
  wrappers in governed.py. Zero module-level bus imports.

4547 tests pass, 4 pre-existing failures, 0 regressions.

+++ packages/acgs-lite/rust/core/src/result.rs
@@ -42,10 +42,13 @@ impl Decision {
                 (severity::DENY_CRITICAL, *rule_idx as i64)
             }
             Decision::Deny { violations, .. } => {
-                let mut bitmask: u64 = 0;
+                let mut bitmask: u128 = 0;
                 for v in violations {
-                    bitmask |= 1u64 << v.rule_idx;
+                    bitmask |= 1u128 << v.rule_idx;
                 }
+                // The PyO3 legacy tuple still exposes an i64 here for backward
+                // compatibility. Bits above 63 truncate at this boundary, while
+                // Python's fallback path still handles large constitutions.
                 (severity::DENY, bitmask as i64)
             }
         }
